+++ interface.go
@@ -10,6 +10,7 @@ import (
 	"net/http"
 	"time"
 
+	"github.com/Conflux-Chain/go-conflux-sdk/middleware"
 	"github.com/Conflux-Chain/go-conflux-sdk/rpc"
 	"github.com/Conflux-Chain/go-conflux-sdk/types"
 	"github.com/ethereum/go-ethereum/common/hexutil"
@@ -36,6 +37,9 @@ type ClientOperator interface {
 
 	CallRPC(result interface{}, method string, args ...interface{}) error
 	BatchCallRPC(b []rpc.BatchElem) error
+
+	UseCallRpcMiddleware(middleware middleware.CallRpcMiddleware)
+	UseBatchCallRpcMiddleware(middleware middleware.BatchCallRpcMiddleware)
 
 	SetAccountManager(accountManager AccountManagerOperator)
 
@@ -137,17 +141,9 @@ type AccountManagerOperator interface {
 
 type RpcRequester interface {
 	Call(resultPtr interface{}, method string, args ...interface{}) error
+	CallContext(ctx context.Context, result interface{}, method string, args ...interface{}) error
 	BatchCall(b []rpc.BatchElem) error
+	BatchCallContext(ctx context.Context, b []rpc.BatchElem) error
 	Subscribe(ctx context.Context, namespace string, channel interface{}, args ...interface{}) (*rpc.ClientSubscription, error)
 	Close()
-}
-
-type CallRPCLogger interface {
-	Info(method string, args []interface{}, result interface{}, duration time.Duration)
-	Error(method string, args []interface{}, resultError error, duration time.Duration)
-}
-
-type BatchCallRPCLogger interface {
-	Info(b []rpc.BatchElem, duration time.Duration)
-	Error(b []rpc.BatchElem, err error, duration time.Duration)
 }